.project/requirements.txt
alembic==1.8.1
APScheduler==3.9.1
base58==2.1.1
click==8.0.4
cryptography==38.0.4
file_read_backwards==2.0.0
Flask_Caching==2.0.1
Flask_JWT_Extended==4.4.3
Flask_Limiter==2.8.1
Flask_SQLAlchemy==2.5.1
ntplib==0.4.0
Pillow==9.3.0
qrcode==7.3.1
requests==2.28.1
SQLAlchemy==1.4.39
SQLAlchemy_serializer==1.4.1
tronpy==0.2.6
Werkzeug==2.0.3
Flask==2.0.3
